+++ Editor/AROcclusionManagerEditor.cs
@@ -5,8 +5,10 @@ using UnityEngine.XR.ARSubsystems;
 namespace UnityEditor.XR.ARFoundation
 {
     [CustomEditor(typeof(AROcclusionManager))]
-    internal class AROcclusionManagerEditor : Editor
+    class AROcclusionManagerEditor : Editor
     {
+        SerializedProperty m_EnvironmentDepthMode;
+        SerializedProperty m_OcclusionPreferenceMode;
         SerializedProperty m_HumanSegmentationStencilMode;
         SerializedProperty m_HumanSegmentationDepthMode;
 
@@ -14,35 +16,64 @@ namespace UnityEditor.XR.ARFoundation
         {
             serializedObject.Update();
 
-            EditorGUILayout.PropertyField(m_HumanSegmentationStencilMode);
-            if (!((HumanSegmentationStencilMode)m_HumanSegmentationStencilMode.enumValueIndex).Enabled())
+            bool isEnvDepthEnabled = ((EnvironmentDepthMode)m_EnvironmentDepthMode.enumValueIndex).Enabled();
+            bool isHumanSegmentationStencilEnabled = ((HumanSegmentationStencilMode)m_HumanSegmentationStencilMode.enumValueIndex).Enabled();
+            bool isHumanSegmentationDepthEnabled = ((HumanSegmentationDepthMode)m_HumanSegmentationDepthMode.enumValueIndex).Enabled();
+            bool isHumanDepthEnabled = isHumanSegmentationStencilEnabled && isHumanSegmentationDepthEnabled;
+
+            EditorGUILayout.LabelField("Environment Depth", EditorStyles.boldLabel);
+            using (new EditorGUI.IndentLevelScope(1))
             {
-                ++EditorGUI.indentLevel;
-                EditorGUILayout.HelpBox("Automatic occlusion during runtime rendering will be disabled with "
-                                        + $"{m_HumanSegmentationStencilMode.displayName} set to "
-                                        + $"{m_HumanSegmentationStencilMode.enumDisplayNames[m_HumanSegmentationStencilMode.enumValueIndex]}.",
-                                        MessageType.Warning);
-                --EditorGUI.indentLevel;
+                EditorGUILayout.PropertyField(m_EnvironmentDepthMode);
+
+                if (!isEnvDepthEnabled && !isHumanDepthEnabled)
+                {
+                    EditorGUILayout.HelpBox("Automatic occlusion during runtime rendering will be disabled with "
+                                            + $"{m_EnvironmentDepthMode.displayName} set to "
+                                            + $"{m_EnvironmentDepthMode.enumDisplayNames[m_EnvironmentDepthMode.enumValueIndex]}.",
+                                            MessageType.Warning);
+                }
             }
 
-            EditorGUILayout.PropertyField(m_HumanSegmentationDepthMode);
-            if (!((HumanSegmentationDepthMode)m_HumanSegmentationDepthMode.enumValueIndex).Enabled())
+            EditorGUILayout.LabelField("Human Segmentation", EditorStyles.boldLabel);
+            using (new EditorGUI.IndentLevelScope(1))
             {
-                ++EditorGUI.indentLevel;
-                EditorGUILayout.HelpBox("Automatic occlusion during runtime rendering will be disabled with "
-                                        + $"{m_HumanSegmentationDepthMode.displayName} set to "
-                                        + $"{m_HumanSegmentationDepthMode.enumDisplayNames[m_HumanSegmentationDepthMode.enumValueIndex]}.",
-                                        MessageType.Warning);
-                --EditorGUI.indentLevel;
+                EditorGUILayout.PropertyField(m_HumanSegmentationStencilMode);
+                if (!isEnvDepthEnabled && !isHumanSegmentationStencilEnabled)
+                {
+                    using (new EditorGUI.IndentLevelScope(1))
+                    {
+                        EditorGUILayout.HelpBox("Automatic occlusion during runtime rendering will be disabled with "
+                                                + $"{m_HumanSegmentationStencilMode.displayName} set to "
+                                                + $"{m_HumanSegmentationStencilMode.enumDisplayNames[m_HumanSegmentationStencilMode.enumValueIndex]}.",
+                            MessageType.Warning);
+                    }
+                }
+
+                EditorGUILayout.PropertyField(m_HumanSegmentationDepthMode);
+                if (!isEnvDepthEnabled && !isHumanSegmentationDepthEnabled)
+                {
+                    using (new EditorGUI.IndentLevelScope(1))
+                    {
+                        EditorGUILayout.HelpBox("Automatic occlusion during runtime rendering will be disabled with "
+                                                + $"{m_HumanSegmentationDepthMode.displayName} set to "
+                                                + $"{m_HumanSegmentationDepthMode.enumDisplayNames[m_HumanSegmentationDepthMode.enumValueIndex]}.",
+                            MessageType.Warning);
+                    }
+                }
             }
+
+            EditorGUILayout.PropertyField(m_OcclusionPreferenceMode);
 
             serializedObject.ApplyModifiedProperties();
         }
 
         void OnEnable()
         {
+            m_EnvironmentDepthMode = serializedObject.FindProperty("m_EnvironmentDepthMode");
             m_HumanSegmentationStencilMode = serializedObject.FindProperty("m_HumanSegmentationStencilMode");
             m_HumanSegmentationDepthMode = serializedObject.FindProperty("m_HumanSegmentationDepthMode");
+            m_OcclusionPreferenceMode = serializedObject.FindProperty("m_OcclusionPreferenceMode");
         }
     }
 }